requirements.txt
pygame-ce
numpy
matplotlib
opencv-python
torch
torchvision